+++ R/TunerGenSA.R
@@ -13,10 +13,10 @@
 #' @inheritSection bbotk::OptimizerGenSA Parameters
 #'
 #' @source
-#' \cite{bbotk}{tsallis_1996}
+#' `r tools::toRd(bibentries[c("tsallis_1996", "xiang_2013")])`
 #'
-#' \cite{bbotk}{xiang_2013}
-#'
+#' @family Tuner
+#' @seealso Package \CRANpkg{mlr3hyperband} for hyperband tuning.
 #' @export
 #' @template example
 TunerGenSA = R6Class("TunerGenSA",