+++ R/TunerNLoptr.R
@@ -21,8 +21,10 @@
 #'
 #' @inheritSection bbotk::OptimizerNLoptr Parameters
 #'
+#' @family Tuner
+#' @seealso Package \CRANpkg{mlr3hyperband} for hyperband tuning.
 #' @source
-#' \cite{bbotk}{johnson_2014}
+#' `r tools::toRd(bibentries["johnson_2014"])`
 #'
 #' @export
 #' @examples